update freindsAdded
Freind list is now updating at homepage
